requirements.txt
openai>=1.0
psycopg2-binary>=2.9
label-studio-sdk>=2.0
Pygments>=2.16
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
scikit-learn>=1.3
